# order-openapi.yaml
openapi: 3.0.1
info:
  title: OpenAPI definition
  version: v0
servers:
  - url: http://localhost:8080
    description: Generated server url
paths:
  /api/v1/orders:
    post:
      tags:
        - order-controller
      operationId: placeOrder
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Order'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                type: integer
                format: int64
  /api/v1/orders/{orderId}:
    get:
      tags:
        - order-controller
      operationId: viewOrder
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/ViewOrder'
components:
  schemas:
    Order:
      type: object
      properties:
        orderId:
          type: integer
          format: int64
        userId:
          type: string
        emailId:
          type: string
        address:
          type: string
        orderDate:
          type: string
          format: date-time
        orderItems:
          type: array
          items:
            $ref: '#/components/schemas/OrderItem'
    OrderItem:
      type: object
      properties:
        orderItemId:
          type: integer
          format: int64
        order:
          $ref: '#/components/schemas/Order'
        bookId:
          type: integer
          format: int64
        quantity:
          type: integer
          format: int32
    ShippingData:
      required:
        - orderId
      type: object
      properties:
        shippingId:
          type: integer
          format: int64
        orderId:
          type: integer
          format: int64
        emailId:
          type: string
        address:
          type: string
        shippingStatus:
          type: string
          enum:
            - SHIPPING_INITIATED
            - SHIPPING_COMPLETE
    ViewOrder:
      type: object
      properties:
        order:
          $ref: '#/components/schemas/Order'
        shippingData:
          $ref: '#/components/schemas/ShippingData'